@@ -12,48 +12,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.6",
-    "@angular/common": "^5.2.6",
-    "@angular/compiler": "^5.2.6",
-    "@angular/core": "^5.2.6",
-    "@angular/forms": "^5.2.6",
-    "@angular/http": "^5.2.6",
-    "@angular/platform-browser": "^5.2.6",
-    "@angular/platform-browser-dynamic": "^5.2.6",
-    "@angular/router": "latest",
-    "@ng-bootstrap/ng-bootstrap": "^1.1.0",
-    "@uirouter/angular": "^1.0.1",
-    "ajv": "6.3.0",
-    "bootstrap": "^4.0.0",
-    "chart.js": "^2.5.0",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^6.0.4",
+    "@angular/cdk": "^6.2.1",
+    "@angular/common": "^6.0.4",
+    "@angular/compiler": "^6.0.4",
+    "@angular/core": "^6.0.4",
+    "@angular/forms": "^6.0.4",
+    "@angular/http": "^6.0.4",
+    "@angular/material": "^6.2.1",
+    "@angular/platform-browser": "^6.0.4",
+    "@angular/platform-browser-dynamic": "^6.0.4",
+    "@angular/router": "^6.0.4",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "@uirouter/angular": "^2.0.0",
+    "ajv": "6.5.0",
+    "bootstrap": "^4.1.1",
+    "chart.js": "^2.7.2",
+    "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "hover.css": "^2.3.0",
+    "hover.css": "^2.3.1",
+    "jasmine": "^3.1.0",
     "jquery": "^3.3.1",
-    "popper.js": "^1.14.1",
-    "rxjs": "^5.5.6",
-    "webpack": "^2.2.0",
-    "zone.js": "^0.8.19"
+    "ng2-cookies": "^1.0.12",
+    "popper.js": "^1.14.3",
+    "rxjs": "^6.2.0",
+    "webpack": "^4.11.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.4",
+    "@angular/language-service": "^6.0.4",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "^10.3.1",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "protractor": "^5.3.2",
+    "ts-node": "~6.1.0",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2"
   }
 }
