@@ -1,13 +1,14 @@
-import { Component } from '@angular/core';
+import {AfterViewInit, Component, Injectable, OnDestroy, OnInit} from '@angular/core';
+import {MatInputModule} from '@angular/material/input';
 import {RecipeService} from './recipe.service';
-
+import {Globals} from './globals';
 @Component({
   selector: 'app-addrecipe',
   templateUrl: '../template/addRecipe.html',
   styleUrls: ['../css/addRecipe.css']
 })
 
-export class AddrecipeComponent {
+export class AddrecipeComponent implements OnInit, OnDestroy {
   name: String;
   categoryID;
   private description: String;
@@ -17,9 +18,16 @@ export class AddrecipeComponent {
   private ingredient_id;
   ingredients;
   categories;
-  constructor(public recipeService: RecipeService) {
+  constructor(public recipeService: RecipeService, public globals: Globals) {
     this.getCategories();
     this.getIngredients();
+  }
+  ngOnInit() {
+   this.globals.isHome = false;
+   this.globals.addrecipe = (true && this.globals.isLoggedIn);
+  }
+  ngOnDestroy() {
+    this.globals.addrecipe = (false && this.globals.isLoggedIn);
   }
   addrecipe(): void {
     this.recipeService.addRecipe(this.name, this.categoryID, this.description, this.instruction, this.picture);
