@@ -1,4 +1,4 @@
-import { Component,Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import {RecipeService} from './recipe.service';
 @Component({
   selector: 'app-recipe',
